menghapus fitur update

+++ app/Views/mahasiswa/index.php
@@ -9,8 +9,7 @@
                         <li class="font-semibold border border-slate-500 px-5 py-2 w-[600px]">
                             <?php echo $mhs['nama'] ?>
                             <span class="inline-block w-1/2 float-right">
-                                <a href="<?php echo BASEURL; ?>/Mahasiswa/hapus/<?php echo $mhs['id'] ?>" class="text-white bg-pink-600 px-3 rounded-full float-right">hapus</a>
-                                <a href="<?php echo BASEURL; ?>/Mahasiswa/ubah/<?php echo $mhs['id'] ?>" class="text-white bg-teal-600 px-3 rounded-full float-right">ubah</a>
+                                <a href="<?php echo BASEURL; ?>/Mahasiswa/hapus/<?php echo $mhs['id'] ?>" class="text-white bg-pink-600 px-3 rounded-full float-right" onclick="return confirm('anda yakin ingin menghapus data ini?')">hapus</a>
                                 <a href="<?php echo BASEURL; ?>/Mahasiswa/detail/<?php echo $mhs['id'] ?>" class="text-white bg-sky-600 px-3 rounded-full float-right">detail</a>
                             </span>
                         </li>
@@ -21,7 +20,7 @@
 
         <!-- tambah modal box -->
 
-        <div class="modal-box w-1/3 bg-white border border-slate-600 rounded-xl py-5 px-10 absolute right-1/2 top-10 translate-x-1/2 shadow-2xl z-50 hidden_modal" id="modal_box">
+        <div class=" modal-box w-1/3 bg-white border border-slate-600 rounded-xl py-5 px-10 absolute right-1/2 top-10 translate-x-1/2 shadow-2xl z-50 hidden_modal" id="modal_box">
             <div class="judul-modal font-semibold text-lg mb-7 flex justify-between">
                 <h1> Tambah Data Mahasiswa </h1>
                 <a href="#" class="hover:cursor-pointer" id="button_modal"><img src="<?php echo BASEURL; ?>/img/close.svg" alt="close"></a>
